revise: remove PostgreSQL feature from core crate
Based on the feedback on #66, I pulled out the PostgreSQL-specific
functionality from the `toasty-core` crate and moved it into the
`toasty-pgsql` crate. I ended up newtyping around
`toasty_core::stmt::Value`, which I like fine for now.
