+++ src/core/src/stmt/value.rs
@@ -261,32 +261,3 @@ impl From<ValueRecord> for Value {
         Value::Record(value)
     }
 }
-
-#[cfg(feature = "postgres")]
-mod postgres_impl {
-    use super::*;
-    use postgres::types::{accepts, private::BytesMut, to_sql_checked, IsNull, ToSql, Type};
-
-    impl ToSql for Value {
-        fn to_sql(
-            &self,
-            ty: &Type,
-            out: &mut BytesMut,
-        ) -> std::result::Result<IsNull, Box<dyn std::error::Error + Sync + Send>>
-        where
-            Self: Sized,
-        {
-            match self {
-                stmt::Value::Bool(value) => value.to_sql(ty, out),
-                stmt::Value::I64(value) => value.to_sql(ty, out),
-                stmt::Value::Id(value) => value.to_string().to_sql(ty, out),
-                stmt::Value::Null => Ok(IsNull::Yes),
-                stmt::Value::String(value) => value.to_sql(ty, out),
-                value => todo!("{:#?}", value),
-            }
-        }
-
-        accepts!(BOOL, INT4, TEXT);
-        to_sql_checked!();
-    }
-}
