+++ src/db/pgsql/src/value.rs
@@ -1,0 +1,34 @@
+use postgres::types::{accepts, private::BytesMut, to_sql_checked, IsNull, ToSql, Type};
+use toasty_core::stmt::{self, Value as CoreValue};
+
+#[derive(Debug)]
+pub struct Value(CoreValue);
+
+impl From<CoreValue> for Value {
+    fn from(value: CoreValue) -> Self {
+        Self(value)
+    }
+}
+
+impl ToSql for Value {
+    fn to_sql(
+        &self,
+        ty: &Type,
+        out: &mut BytesMut,
+    ) -> std::result::Result<IsNull, Box<dyn std::error::Error + Sync + Send>>
+    where
+        Self: Sized,
+    {
+        match &self.0 {
+            stmt::Value::Bool(value) => value.to_sql(ty, out),
+            stmt::Value::I64(value) => value.to_sql(ty, out),
+            stmt::Value::Id(value) => value.to_string().to_sql(ty, out),
+            stmt::Value::Null => Ok(IsNull::Yes),
+            stmt::Value::String(value) => value.to_sql(ty, out),
+            value => todo!("{:#?}", value),
+        }
+    }
+
+    accepts!(BOOL, INT4, TEXT);
+    to_sql_checked!();
+}
